Fix NVBench Warnings

+++ nvbench/detail/l2flush.cuh
@@ -35,7 +35,7 @@ struct l2flush
     if (m_l2_size > 0)
     {
       void *buffer = m_l2_buffer;
-      NVBENCH_CUDA_CALL(cudaMalloc(&buffer, m_l2_size));
+      NVBENCH_CUDA_CALL(cudaMalloc(&buffer, static_cast<std::size_t>(m_l2_size)));
       m_l2_buffer = reinterpret_cast<int *>(buffer);
     }
   }
@@ -52,7 +52,8 @@ struct l2flush
   {
     if (m_l2_size > 0)
     {
-      NVBENCH_CUDA_CALL(cudaMemsetAsync(m_l2_buffer, 0, m_l2_size, stream));
+      NVBENCH_CUDA_CALL(
+        cudaMemsetAsync(m_l2_buffer, 0, static_cast<std::size_t>(m_l2_size), stream));
     }
   }
